Refactor and Modularize
- Modularize code based on Github API v3 structure.
- Run `cargo clippy`
- Run `cargo fmt`
